requirements.txt
sqlalchemy 
psycopg2-binary 
python-dotenv